chore: moved LHD profile getters and types
