+++ src/state/lhd/hooks/useGetLHDProfile.ts
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query'
 import axios from 'axios'
 
 // Types
-import { TokenProfile } from 'utils/types/lhd'
+import { TokenProfile } from '../types'
 
 // Constants
 import { LHD_API_TEMP } from 'config/constants/api'
