+++ src/views/LHD/components/FullProfile/components/InfoCards/index.tsx
@@ -2,7 +2,7 @@ import { Flex, Text } from 'components/uikit'
 import HealthSummaryRow from './HealthSummaryRow'
 import { formatDollar, formatValue } from 'utils/formatNumbers'
 import { useTranslation } from 'contexts/Localization'
-import { chartExtras, LiquidityOwner, LiquidityPool, TokenProfile } from 'utils/types/lhd'
+import { chartExtras, LiquidityOwner, LiquidityPool, TokenProfile } from 'state/lhd/types'
 import { styles } from './styles'
 import dynamic from 'next/dynamic'
 import { Svg } from 'components/uikit'
