+++ src/views/LHD/components/FullProfile/components/TopSectionCards/index.tsx
@@ -5,7 +5,7 @@ import IconButton from '../IconButton'
 import ChainsIcons from '../ChainsIcons'
 import ProgressBar from '../../../ProgressBar'
 import { getColor } from '../../../../utils/getColor'
-import { ExternalDataOption, TokenProfile, TokenProfileLinks } from 'utils/types/lhd'
+import { ExternalDataOption, TokenProfile, TokenProfileLinks } from 'state/lhd/types'
 import useModal from 'hooks/useModal'
 import SharableCard from '../../../SharableCard'
 import { styles } from './styles'
