+++ src/views/LHD/components/TokensProfileList/index.tsx
@@ -17,7 +17,7 @@ import SearchBar from '../SearchBar'
 import FilterModal from '../SearchBar/FilterModal'
 
 // Types
-import { LHDProfiles, Filters } from 'utils/types/lhd'
+import { LHDProfiles, Filters } from 'state/lhd/types'
 
 interface TokensProfileListProps {
   simpleProfiles: LHDProfiles
@@ -29,13 +29,13 @@ interface TokensProfileListProps {
 }
 
 const TokensProfileList = ({
-                             simpleProfiles,
-                             isLoading,
-                             isSearchQuery,
-                             appliedFilters,
-                             setIsSearchQuery,
-                             handleFiltersChange,
-                           }: TokensProfileListProps) => {
+  simpleProfiles,
+  isLoading,
+  isSearchQuery,
+  appliedFilters,
+  setIsSearchQuery,
+  handleFiltersChange,
+}: TokensProfileListProps) => {
   // TODO: Come back to double check the sorting and add types/enums
   const [sortCol, setSortCol] = useState(appliedFilters.sort ? 'Market Cap' : 'Score')
   const [sortType, setSortType] = useState<'asc' | 'desc'>('desc')
@@ -74,9 +74,7 @@ const TokensProfileList = ({
 
   const renderSortedProfiles = useMemo(() => {
     return sortProfiles(simpleProfiles.data, sortCol, sortType)?.map((simpleProfile, index) => {
-      return (
-        <TableRow key={`simpleProfile${index}`} index={index} simpleProfile={simpleProfile} />
-      )
+      return <TableRow key={`simpleProfile${index}`} index={index} simpleProfile={simpleProfile} />
     })
   }, [simpleProfiles.data, sortCol, sortType])
 
@@ -115,7 +113,7 @@ const TokensProfileList = ({
               p: '20px',
             }}
           >
-            <Svg icon='placeholderMonkey' />
+            <Svg icon="placeholderMonkey" />
             <Text sx={{ fontSize: '12px', fontWeight: 500, color: 'textDisabled' }}>{t('No Results Found')}</Text>
           </Flex>
         )}
